Clip video rect to physical screen dimensions
The CVideoPlayerUtility API requires both a native window handle
and an absolute screen rectangle in order to define the location
of the rendered video output.  On certain devices, such as the
Nokia E75, which runs S60 3.2, if the absolute rectangle extends
outside the physical screen extent, no video is rendered.

This change works around this defect in the platform by clipping
the video rectangle to the physical screen extent.

Task-number: QTBUG-5467
Reviewed-by: trustme