Optimize QObjectPrivate::isSignalConnected
It can be seen in the GraphicsView benchmark.
This function has to be as fast as possible.
Make the function inline.

We do not need 64bit for the connectedSignals bit array because
it is very unlikely that an object will have more than 32 signals.
(In Qt 4.5 it could have hapenned as the slot were counting in
the index)

Reviewed-by: bnilsen
Reviewed-by: Brad