Fixed context menu placement partially outside screen bug in Symbian.
In style demo example when spinbox was long tapped, the context menu
was placed partially outside screen. The preconditions for bug were:

1. Tap position is near screen right edge
2. Minimum space width requested by context menu is > sreen width

The new (simplified) logic is as follows:

1. If menus right edge in requested position is outside screen right
edge, the menu is moved to left enough to fit on screen. i.e.
    if (tap position + menu width > screen right )
        x = screen right - menu width

2. As a result of above statement the x position might go outside left
boundary of screen. This is fixed by next statements
    if( x < screen left )
        x = screen left

Task-number: QTBUG-6164
Reviewed-by: Alessandro Portale