Fix QFontDialog::getFont on Mac OS X when using an invalid initial font
If you passed in a QFont with a family name that did not resolve to
QFontDialog::getFont(), it would not select any font in the panel,
and it would always return the default font, regardless of what you
actually selected in the dialog. This was because it would try to
resolve the requested family name, rather than the actual family name
of the initial font. That in turn caused the NSFont* returned by the
system to be null, which, when set on the font manager, caused the
manager to always return 0 for selectedFont.

Task-number: QTBUG-6071
Reviewed-by: Trond