Compile webkit for win32-icc
The reason for the first hunk is that the Intel compiler does not
support the __has_trivial_constructor type trait.
The Intel Compiler can report itself as _MSC_VER >= 1400. The reason
for that is that the Intel Compiler depends on the Microsoft Platform
SDK, and in order to try to be "fully" MS compatible it will "pretend"
to be the same MS compiler as was shipped with the MS PSDK. (Thus,
compiling with win32-icc with VC8 SDK will make the source code "think"
the compiler at hand supports this type trait).

The second hunk is simply for including stdint.h since MS does not
ship that in their PSDK.

Reviewed-by: Simon Hausmann