X11: Avoid loading all fonts in system for text with line breaks
The line separator code point has no glyph in any font, something we
would detect by running through a complete font merging loop and trying
all fonts in the system on X11. This would cause the memory consumption
to sky rocket the first time you added text with line breaks to an
application, and it was an unnecessary performance hit. Now we special
case the line separator and do no attempt to search for it in the
fonts.

Task-number: QTBUG-4537
Reviewed-by: Simon Hausmann