Creator crashes when reloading externally modified .ui files on Mac
On Mac QWidget::destroy() sends an AcceptDropsChange event after
clearing the guards for QPointer. This was used to store a QPointer to
the widget being deleted & that will never be cleared.
The fix removed the setAcceptDrops() from destroy. And as an extra
protection make sure designer will not treat that event as interesting.

Task-number: QTCREATORBUG-307
Reviewed-by: Denis Dzyubenko
Reviewed-by: Friedemann Kleint