Fixed bug which caused Phonon backend error messages to be suppressed
When the mediaplayer receives a state change into the ErrorState, it
calls pause() on the media object.  Previously, this caused the
backend to transition into PausedState.  When the mediaplayer subsequently
called errorString() to retrieve the error message, an empty string was
returned because the backend was no longer in the ErrorState.

Task-number: QTBUG-4994
Reviewed-by: trustme
(cherry picked from commit dfa9343cc1e634eebdb0f3a2cf931ada9829ae6b)