Changed call sequence of seeking in Phonon MMF backend, for streaming
Modified the sequence of calls made to the MMF APIs when seeking
during ongoing playback.  This fixes a bug found during early testing
of streaming playback, whereby playback would not resume following
the seeking operation.  This was due to an interaction between the
pause / seek / play operations, and the buffering callbacks received
from the MMF, which caused the backend to enter an incorrect state.

Task-number: QTBUG-4660
Reviewed-by: Frans Englich