Reduce double-copying of textures when flipping upside down
bindTexture() flipped images in-place, to reduce data copying.
But there is one case where the in-place is worse: when the
QImage is not detached.  In that case, the flip was copying
the entire image and then flipping the lines, effectively
processing the contents twice.  The new version uses mirrored()
to reduce the overhead for non-detached images.

Reviewed-by: Samuel