Added a placeholderText(*) property to QLineEdit.
Please note: Qt for Maemo5 will already use it in 4.6, while the other
platforms will be enabled in 4.7 (otherwise be would break BIC there)

This commit adds a so-called placeholder text for line edits. If the widget
doesn't have focus and the text() is empty, this placeholder will be shown.

(*) also known as hint, click-message or descriptive text

Reviewed-by: jasplin