Cocoa: refactor common code into a shared file
QCocoaWindow and QCocoaPanel has many methods that are implemented
_exactly_ the same way. Those methods are just copied/pasted between
the classes because of inheritance problems. This is error-prone as
bugs tends to be fixed inside one of the classes, but easily forgotten
in the other class. This patch refactors out this code into a new
file that is simply #included from the two classes.
We do this fix for a patch release to ease a couple of fixes that is about
to be integrated.

Reviewed-by: Prasanth
Reviewed-by: Denis