QS60Style: Always store changed theme palette
When style updates its background brush, it updates QApplication's
palette with new background. Also member variable m_backgroundBrush is
updated.

What is NOT updated is the member variable m_themePalette, which should
also contain correct updated palette.

This change updates m_themePalette and moves palette updations to occur
in one place so that in all use-cases the palettes are modified in a
unified way.

Task-number: QTBUG-6427
Reviewed-by: Alessandro Portale