Cocoa: impl. performance boost for dnd
It turns out that registering drag types for each NSView that can
receive drop events is _really_ slow. And many widget in Qt
subscribe for DnD (QTextEdit, QScrollArea, etc), so the result is
an application that will spend startup time preparing for DnD. For
some edge cases, we're talking several seconds!
This patch removes this overhead by moving drag type registering
out of NSView, and into NSWindow (that is, QCocoaWindow and
QCocoaPanel).

Task-number: QT-1586
Reviewed-by: Prasanth