Made sure that we didn't reference dangling parallel anchors.
We cannot delete the parallel anchors before we do the actual vertex
restoration, since the parallel anchors created by vertex simplification
will always be connected to some of the AnchorVertexPair found in
simplifiedVertices that we iterate over later.

Thus, we must delete the parallel anchors created by vertex
simplification after we have done the restoration of the vertices. (the
parallel anchor will be properly dealt with (i.e.  taken out of the
graph as we do the restoration)), so not deleting it won't cause it to
be revisited later if that is a worry.

Reviewed-by: Eduardo M. Fleury