Switched S60 QDesktopServices implementation to CDocumentHandler based.
RApaLsSession based implementation did not succeed to start music
player in some S60 devices to playing mode. The only found solution to
this problem was to start using CDocumentHandler based implementation.
CDocumentHandler adds a new S60 dependency, but it is needed to make
QDesktopServices APIs fully fuonctional. However if Qt is comfigured
without S60 we still fallback to RApaLsSession implementation.

With CDocumentHandler the files are also opened as embedded, due to the
fact that swicthing files in stand-alone mode would require SwEvent
capability.

Task-number: QTBUG-4699
Reviewed-by: Miikka Heikkinen