Added auto test for the cursor position when doing block undo and redo
Block changes in different selections happens quite frequently with
Qt Creator, e.g. when renaming local variables. This test tests the
desired cursor position behaviour: after an undo, the cursor should be
where it was when the block operation started. After a redo, the cursor
should be where it was after the block operation ended.