Fix cursor positiong after block undo and redo
Block changes in different selections happens quite frequently with
Qt Creator, e.g. when renaming local variables. The desired behaviour
which the tst_QTextCursor::cursorPositionWIthBLockUndoAndRedo() tests,
is like this: after an undo, the cursor should be where it was when
the block operation started. After a redo, the cursor should be
where it was after the block operation ended.

Reviewed-by: Simon Hausmann