QS60Style: Overwrites correct stylehint color values
QS60Style overwrites correct theme colors with QCommonStyle values
in QS60Style::styleHint(). In that method returned value is initialized
as -1 and then if the value is still -1 after switch statement for
different stylehints, the stylehint is queried from common style.

Problem is that it is also getting RGB values set into return value.
White color as HEX is FFFFFF and -1 as HEX is FFFFFF.

So, if theme color is white for Group Text title, or for Grid lines,
color is asked from QCommonStyle instead.

As a fix, in 'default' case ask from QCommonStyle. Also, use RGBA
instead of RGB values.

Reviewed-by: Alessandro Portale