Make sure that QtGui links with QtCore for the win32-icc mkspec.
Fix it by exporting QObjectData.  Since this is a compiler bug we will only
export it for the compiler that has this problem. (Intel Compiler on Windows)
The Intel compiler seems to inline the destructor of QObject (or at least
parts of it) even if it is declared virtual. Thus, it seems that the
destruction of the QScopedPointer<QObjectData> member is inlined into QtGui,
and that code will eventually call the destructor of QObjectData, and since
QObjectData was not exported from QtCore it would crash.

Reviewed-by: Olivier Goffart