network API: add support for HTTP multipart messages
This commit adds two new classes, QHttpPart and QHttpMultiPart, and
two new overloads to QNetworkAccessManager:
post(const QNetworkRequest &request, QHttpMultiPart *multiPart) and
put(const QNetworkRequest &request, QHttpMultiPart *multiPart).

With those classes, it is possible to do a HTTP POST with a multipart
message in a memory-saving way: The data from the parts is not copied
when read from a file or another QIODevice.

Reviewed-by: Markus Goetz
Task-number: QTBUG-6222