Automatically destroy VG pixmaps when the last window surface goes away
Under S60, Qt will destroy the window surfaces of an application that
goes into the background, which frees up EGL surface objects.  But the
VGImage's for pixmaps, and the EGLContext, were still using GPU memory.

This change keeps track of the number of widgets / window surfaces that
are in use and then calls hibernate() on all QVGPixmapData objects when
it goes to zero.  Once all the VGImage's are destroyed, the EGLContext
should also be destroyed.

Task-number: QT-2555
Reviewed-by: Sarah Smith