Rewrite Positioners
Cleans up code, reduces memory footprint and increases performance.
Largely by removing large chunks of functionality that we don't
actually use.

Also now that we are clearer on what the new positioner transitions are
(how the item moves it's position when it is positioned by a positioner)
a remove transition makes no sense and is discarded, and the auto tests
have been updated as add transitions are slightly different.