Added dimming support for disabled softkeys in Symbian.
If QAction::setEnabled(false) is called, the CBA buttons are dimmed
to have visual indication about disabled state.

Since enabled/disabled state of buttons in QDialogButtonBox is
controlled via QPushButton::setEnabled API, and because button box
content in Symbian is mapped to sofkkeys we also need to have proxy
for button enabled state to forward the information for underlying
QAction.

Reviewed-by: Sami Merila