Cannot drag scrollbars in a QPlainTextEdit on Cocoa.
The QPlainTextEdit can change the scroll ranges while dragging the
scrollbar. This will eventualy call QWidget::raise(), on Cocoa it was
done by removing the NSView and adding it back. This causes problems
like resetting internal state while a mouseDragged was active on the
view. The fix we will now sort the views based on their Qt-z-order.
lower() & stackUnder() also fixed like this.

Reviewed-by: Denis