Fix GL_BGRA formats under OpenGL/ES systems
OpenGL/ES requires that the internal and external formats be the
same when calling glTexImage2D().  This caused problems with
devices that had the GL_EXT_bgra extension.

This change makes the formats the same just before the upload and
also makes sure that the pixel type is GL_UNSIGNED_BYTE when
GL_BGRA is used.  No change for desktop systems.

Reviewed-by: Donald Carr