Minor fixes to softkey dimming support (commit 245c9cc0).
In Qt invisible actions are disabled by default, and our "Options"
softkey is set to invisible in order that it is not show in context menu.
Thus we need don't want to dim in softkey even it is disabled.

Additionally, QDialogButtonEnabledProxy need to set the initial enabled
state for proxy action from button.

Fixes bugs in commit: 245c9cc0

Reviewed-by: TrustMe