Use realpath only on OS we know it works on.
As far as we know realpath(X,0) extenstion is only supported on Linux, Mac OS X
starting with 10.6 and on Symbian.

Here goes the trick: realpath() on Mac properly handles file systems
case-sensitivity, meaning two files with different cases will are the same file
if the file system is case insensitive (which is the default on Mac). However
the QFSFileEngine will still say that the file system is case sensitive because
on Mac you can have several drives with different file systems (with different
case sensitivity), and QFSFileEngine doesn't allow to return different values
depending on the file path, so we still say that the file system is
case-sensitive, which is the safiest behavior.

This changes the behavior on Mac, but changes it to be correct.

Reviewed-by: Markus Goetz
Reviewed-by: João Abecasis