Fix possible off-by-one inconsistency against system look and feel when drawing vertically aligned text
When text is drawn in a font which has an odd-numbered height and it's
vertically aligned in a rectangle which has an even-numbered height
(or vice versa) we have to round the y-position, as the backend cannot
position the text in subpixels. The y-position in this case will be at
half a pixel and we can either round it up or down without increasing
the size of the error.

In Qt 4.5, the font height would be reported as one higher than what it
actually was. Therefore, in many widgets such as comboboxes, we would
appear to round the y-position down when faced with this issue. This is
apparently also what Windows does, probably because most latin fonts
have more glyphs extending into the maximum descent than the ascent.

In Qt 4.6, the font height bug was fixed, and since we use qRound() we
now round half-pixels up, thus moving all the text faced with this issue
one pixel down. This looks odd and inconsistent, both with the platform
look and feel and with older versions of Qt.

The patch reverts to the old look and feel by rounding the half pixels
down rather than up for the particular case in question.

Task-number: QTBUG-6578
Reviewed-by: Samuel