Rewrote most of the regular pointer handling.
The old implementation had been hacked on for a while and needed
cleanup. The new code is heavily based on looking at the behavior of
other platforms. It also reuses more of the cross platform code,
which improves the handling of Enter and Leave events.

We also switched to letting Symbian grab the pointer automatically
when pressing down the mouse button, which improves things
considerably compared to doing it ourselves.

Popups should also work a lot better after this fix, since they were
not really handled at all in the old code.

The old code had calls to set the Symbian cursor sprite. This code
has been removed since that code is now being called from within
dispatchEnterLeaveEvents().

In addition, there was code to check whether the up key event had
been left out by the platform. This was solved a bit differently now,
instead putting the code in the section that handles virtual mouse,
since that is where the problem occurs.

Task:     QTBUG-4990
RevBy:    Shane Kearns
AutoTest: N/A, Platform specific code that an autotest cannot catch.
          Lots of manual testing was done on normal examples as well
          drag'n'drop examples and it seemed to work fine.