Use realpath() only on systems we know it works on.
We use realpath(X,0) extension that is defined by the latest POSIX standard and
not many systems support it at the moment. For now we limit it to Linux and
Symbian. Mac supports it starting with 10.6, and we'll implement it properly
for Mac in 4.7. We know that neither *BSD systems nor Solaris do not support
it.

Reviewed-by: Markus Goetz