QS60Style: Do not let QCommonStyle draw PE_PanelScrollAreaCorner
If QS60Style lets QCommonStyle to draw PE_PanelScrollAreaCorner,
common style draws transparent small rect to the end of scrollbar.
This is due to our style's palette's Base is Qt::Transparent.

Error is only visible when two dialogs/windows are on top of each other
and their background is such that a hole in the upper dialog's
background will make the hole "visible"
(i.e. background is not similar).

As a fix, QS60Style should not let QCommonStyle draw the primitive,
but instead handle itself (and do nothing with it).

Task-number: QTBUG-6657
Reviewed-by: Alessandro Portale