Assistant: Make doc file watching work for the intended use case.
The QFileSystemWatcher signals the changing of a file several times,
with the first one coming too early. As a result, we are likely
checking a file that is not yet fully constructed.
The workaround/solution is to delay reacting to the change until
after the (hopefully) last signal from the QFileSystemWatcher.

Reviewed-by: kh1