Prevent posted event starvation by native Windows messages
Avoid live-lock if native messages are generated quickly enough by
allowing posted events to be sent if more than 10ms has passed since
we last sent them. This means that we can allow multiple calls to
sendPostedEvents() during a single call to processEvents(), but only
when the QEventLoop::EventLoopExec flag is passed to processEvents().

Task-number: QTBUG-1697