add {QString,QByteArray,QList,QLinkedList,QHash,QMap,QVector}::isSharedWith()
these functions just compare the d pointers of two objects. this can be
used to verify the validity of cached data: you keep a copy of the
original data in your cache. consequently, any changes to the original
data must detach, so this function reflects whether the cached data is
still up-to-date.
of course, this is not thread safe, as the data might change right after
you check it, but that's a general problem of cache coherency which
needs to be handled further up the stack.

the functions are internal because the other detaching-related functions
are internal as well for reasons beyond me.

Reviewed-by: brad