Prevented leak of keys in QPixmapCache.
Removing a pixmap from the pixmap cache using the new QPixmapCache::Key
API left the keys dangling in the QCache's internal QHash. The problem
is that the Key is invalidated as soon as the QPixmapCacheEntry is
destroyed, thus removing it from the hash failed. Reordering the
destruction of the object and the removal of the key in QHash fixes the
problem.

Reviewed-by: Alexis
Reviewed-by: Thiago