Fixed qgraphicseffectsource autotest.
Prevent crashing when asking for a source pixmap in logical coordinates
and PadToEffectiveBoundingRect mode. Also, change the expected values
for the test that uses LogicalCoordinates and
PadToEffectiveBoundingRect mode to reflect that the boundingRectFor()
function applies padding in device coordinates (and in this case there
is a scale by 2 when drawing the effect).

Reviewed-by: Bjørn Erik Nilsen