Slow spinbox on N95 when using keys Up/Down
In the style demo, spinbox widget, we noticed that to go from 0 to 99
, with key Up, takes 35 seconds. This was too slow.

It turned out that no acceleration has been implementd when using
keyboard, although acceleration works when mouse used.

We fixed this by using timer events, as it is the case for mouse events.
We also needed to get what is the keyboard repat rate interval from system.
This value is used to set timer intervals in more natural way.

Task-number: QT-927
Reviewed-by: janarve