Focus frames in mac style gets clipped.
The focus frame on mac is a separate widget. The parent of the focus
frame is same as the parent of the widget who has the focus. So if
there is not enough room for a focus frame between the parent and
the widget, it gets clipped to the parent.

This parent is now changed to the nearest (in hierarchy) top level,
toolbar or scrollarea-viewport. A clip area is set on the focus
frame widget based on the visible region of the widget. This
reduces the chances of having clipped focus frames. It is always
better to leave enough margin for widgets which shows focus frames.

Task-number: 6567
Reviewed-by: Richard Moe Gustavsen
Reviewed-by: Jan-Arve