FEP indicator shown in status pane when it should not
FEP indicator is shown in few cases in statusbar when it should not be.
Case a)
Application contains item view.
QAbstractItemView is initialized as having WA_InputMethodEnabled on,
irregardless of functionality of each separate cell. So we need
to check the model for the itemview cell about to receive focus,
is it editable. If it isn't then, we'll set the WA_InputMethodEnabled
as false.
This will fix the FEP indicators in a lot of cases: lists with cells,
empty lists, combobox lists, ...

Case b)
Combobox also initializes itself with WA_InputMethodEnabled active.
Even if the fix a) fixed the list inside combobox, the widget itself
when receiving focus needs also fixing. If we check the editable
flag of combobox when initializing the comment, we can set the
flag off.

Task-number: QTBUG-5705
Reviewed-by: axis