Fix vertical text centering on Mac OS X
Change 04d18b38 fixed the windows and ft font engines so that the height
reported in QFontMetrics is now correct, but this make vertical text
centering one off as the platform defaults is to round down rather than
up when the vertical position of the text is in the middle of a pixel.
When this was fixed in change 1de8a5b, the vertical centering broke on
Mac, since it still reported a too great font height. This patch applies
the same hack to the mac font engines to report the correct font
height.

Task-number: QTBUG-6770
Reviewed-by: joerg