Text cursor is not shown properly in line edits on Mac OS X.
The mac stlye disables the blinking cursor when there is a selection in
the line edit. The cursor is not shown back when this selection is
removed (either by mouse or key press events).
This patch watches the selection changed signal from the line control
and updates the cursor status accordingly.

Task-number: QTBUG-6378
Reviewed-by: Olivier Goffart