Optimize QThreadStorage with QVector
QThreadStorageData::get might be accessed very often. QMap lookup
has proven to be too slow.

We can expect about 20 instances of QThreadStorage in a typical
applications (tested with some KDE applications)
So a QVector is more suited

Note: we now re-use the ids. Which means that if the QThreadStorage
is destroyed before the QThread, and another one is created,
we might get crashes (instead of a warning printed to the console)

Reviewed-by: brad