Prevent access to non-existent memory in triagulating stroker
In the triangulating stroker, the last point was being duplicated in
dashed paths.  But because QDataBuffer::add() takes a ref to a float
rather than a float, it would resize the data buffer and then try to
fetch the values out of a pointer to the original buffer memory.

This change copies the values into temporary variables before
resizing the array.

Task-number: QTBUG-6045
Reviewed-by: Sarah Smith