Tool for launching symbian apps on the phone from windows command line
The tool uses TRK to launch the application.
TRK is a debug agent, available as a signed package for consumer phones.
For Nokia phones, it is distributed with carbide; for other manufacturers
it can be downloaded from their developer websites.
The launcher code is reused from Qt creator, with a patch to allow us
to pass command line arguments to the process being launched.

The "make run" target is enhanced to support running on target as well as
the emulator. Like the "make sis" target, the last platform to have been
built is the one that will be launched.

The runonphone tool needs to be built in a Qt environment configured for
windows, and requires at least windows XP (Symbian development requires
windows XP SP2). Current proposal is to include a statically linked exe
in the bin directory for binary packages.

Reviewed-by: Janne Koskinen