Remove unused legacy code from the jpeg image plugin
A long time ago, in a Qt version far, far, away there was a
parameter string mechanism for asking for header details,
requesting scaling, and so on.  This has since been replaced
with actual real API's and it is no longer possible to pass
such parameter strings to the image plugins.  This change
removes the crufty beloved old code.

Reviewed-by: Sarah Smith