Deal with test cases that crash or hang
Added an optional timeout to runonphone - the application will be killed
after this time. Used when autotesting unattended, as some tests can hang.

Handled the just in time debug halting the application when it is about
to crash, by terminating the application. In future, we could capture a
call stack or something here.

Also added quiet/verbose options to control the amount of output from
runonphone.

Reviewed-by: Janne Koskinen