Setting background color to a QDialog doesn't work
Since dialogs use their own theme graphics for background, regular
method of checking the palette fails (as generic background brush is
probably unchanged). Therefore to check if the dialog background is
using default (=theme graphic), it needs to check the cache key value
of background - so that it matches with the one in stored theme
palette.

Task-number: QTBUG-5898
Reviewed-by: Janne Koskinen