Fixes crash when widget with WA_StaticContents child become toplevel.
Happens for example if a DockWidget is undocked and has a child
whith the WA_StaticContents attribute.

The parent does not change (so newParent is false) but still, the
top level widget change. So staticWidget need to be moved to the
new backingstore.

Reviewed-by: Benjamin Poulain
Task-number: QTBUG-6883
(cherry picked from commit 03baad51d584778a99f4b393232ccc1312db3391)