Remove superfluous enable/disable vertex arrtib arrays
Now that the shader manager takes care of enabling/disabling the
vertex attribute arrays for us, the GL2 paint engine doesn't have
to do it. This reduces GL state changes within the paint engine
and provides significant performance improvements. For a given test
case (25,000 3x3px solid rects), the improvement is 67% on desktop
(nVidia) and 9% on embedded (SGX).

Reviewed-By: Kim