Fixes painting artifacts when using CacheBackground in a QGraphicsView.
The problem was that when the background cache was invalidated, it was
entirely recreated but only the exposed area of the view was repainted
in it, causing the cache to be partly empty in some cases. Now the
background cache is always fully repainted when it is invalidated.

Task-number: QTBUG-6935
Reviewed-by: ogoffart
(cherry picked from commit 2600fd42117913b427d07e510724b0ea5e355205)