Fixed QGLWidget::renderText().
Fall back and use the GL 1 engine for the renderText() functions.
Getting it to work with the GL 2 engine is a futile effort.
Making it work with renderPixmap() in the GL 2 engine is not possible
at all, since software contexts in general do not support shader
programs.

Task-number: QTBUG-5002, QTBUG-6931
Reviewed-by: Kim