Change the QtDBus watched service list to a map to its owner.
Instead of simply counting how many times we're watching a service,
also map a service to its owner. This allows us to provide a faster,
cached GetNameOwner too.

Technically, this could mean that a service that is started in-between
event loops could remain invalid for a longer period of time. I'm not
sure this is an issue -- will need some testing in KDE.

Task-number: QTBUG-5979
Reviewed-By: TrustMe