New lance test for checking aliased vs antialiased rendering
This test renders various primitives with anti-aliasing both on and
off. It then repeats the render at several non-integer offsets.

Note: The reference image was generated with the raster engine,
which seems to have off-by-one errors on fills when rendering
aliased (It seems to ceil the coords rather than round them).

Reviewed-By: Trustme