Introduce new "snapToPixelGrid" flag to GL2 engine for drawText
When we're rendering text, the glyphs need to be aligned to the
pixel grid otherwise we get strange artifacts. Normally text is
drawn at integer coordinates, however it is still possible to
have a transform which translates by a non-integer offset. This
patch adds a flag to the engine which can be used to snap any
translate to the pixel grid.

Task-number: QTBUG-7094
Reviewed-By: Kim