QScript: Lookup the native setter from the prototype
Usefull if we have 'native' properties with setter in a prototype
This happen if you use a QObject wrapper as a prototype.

Use getPropertyDescriptor that look up the prototype in order to
know if we have a setter.

Note that we cannot relly on PropertDescriptor::isAccessorDescriptor
as the Getter or Setter attributes are not necesserly updated correctly
when updating properties. (See the workaround QScriptValuePrivate::propertyFlags,
and tst_QScriptValue::getSetProperty with object7)

Task-number: QTBUG-5749  (also need the previous patch)
Reviewed-by: Kent Hansen