Fixed a bug with distribution of spans.
If a span required more size than the rows/columns it spanned, the
size of the span was not distributed to the rows it spanned. The
result was that the size hints of the layout was not correct,
causing the layout to be potentially smaller than the spanning item.

Task: QT-2261
Reviewed-by: Alexis